pyproject.toml
[project]
name = "remesh"
version = "0.1.0"
description = ""
authors = [
    {name = "Reece Humphreys",email = "reecewh@icloud.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "trimesh (>=4.6.6,<5.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
